# Log additional performance metrics in `create_snapshots_reports_scorecard.py`